requirements.txt
Django>=1.9
djangorestframework>=3.3
django-oauth-toolkit
pillow

# Optional
Werkzeug
django-extensions
